+++ ajax/timeline.php
@@ -34,68 +34,70 @@ header("Content-Type: text/html; charset=UTF-8");
 
 Session::checkLoginUser();
 
-if (!isset($_POST['action'])) {
+if (!isset($_REQUEST['action'])) {
    exit;
 }
 
-switch ($_POST['action']) {
+switch ($_REQUEST['action']) {
    case "change_task_state":
-      if (!isset($_POST['tasks_id'])) {
+      if (!isset($_REQUEST['tasks_id'])) {
          exit();
       }
       $task = new TicketTask;
-      $task->getFromDB(intval($_POST['tasks_id']));
+      $task->getFromDB(intval($_REQUEST['tasks_id']));
       if (!in_array($task->fields['state'], [0, Planning::INFO])) {
          echo $new_state = ($task->fields['state'] == Planning::DONE)
                               ? Planning::TODO
                               : Planning::DONE;
-         $task->update(['id'         => intval($_POST['tasks_id']),
-                             'tickets_id' => intval($_POST['tickets_id']),
+         $task->update(['id'         => intval($_REQUEST['tasks_id']),
+                             'tickets_id' => intval($_REQUEST['tickets_id']),
                              'state'      => $new_state]);
       }
       break;
    case "viewsubitem":
       Html::header_nocache();
-      if (!isset($_POST['type'])) {
+      if (!isset($_REQUEST['type'])) {
          exit();
       }
-      if (!isset($_POST['parenttype'])) {
+      if (!isset($_REQUEST['parenttype'])) {
          exit();
       }
 
-      if (($item = getItemForItemtype($_POST['type']))
-          && ($parent = getItemForItemtype($_POST['parenttype']))) {
-         if (isset($_POST[$parent->getForeignKeyField()])
-             && isset($_POST["id"])
-             && $parent->getFromDB($_POST[$parent->getForeignKeyField()])) {
+      $item = getItemForItemtype($_REQUEST['type']);
+      $parent = getItemForItemtype($_REQUEST['parenttype']);
 
-            $ol = ObjectLock::isLocked( $_POST['parenttype'], $parent->getID() );
-            if ($ol && (Session::getLoginUserID() != $ol->fields['users_id'])) {
-               ObjectLock::setReadOnlyProfile( );
-            }
-
-            Ticket::showSubForm($item, $_POST["id"], ['parent' => $parent,
-                                                                        'tickets_id' => $_POST["tickets_id"]]);
-         } else {
-            echo __('Access denied');
-         }
-      } else if ($_POST['type'] == "Solution") {
-         $ticket = new Ticket;
-         $ticket->getFromDB($_POST["tickets_id"]);
+      if ($_REQUEST['type'] == "Solution") {
+         $parent->getFromDB($_REQUEST[$parent->getForeignKeyField()]);
 
          if (!isset($_REQUEST['load_kb_sol'])) {
             $_REQUEST['load_kb_sol'] = 0;
          }
-         $ticket->showSolutionForm($_REQUEST['load_kb_sol']);
 
-         // show approbation form on top when ticket is solved
-         if ($ticket->fields["status"] == CommonITILObject::SOLVED) {
-            echo "<div class='approbation_form'>";
-            $followup_obj = new TicketFollowup();
-            $followup_obj->showApprobationForm($ticket);
-            echo "</div>";
+         $sol_params = [
+            'item'         => $parent,
+            'kb_id_toload' => $_REQUEST['load_kb_sol']
+         ];
+
+         $solution = new ITILSolution();
+         if (isset($_REQUEST['id']) && (int)$_REQUEST['id'] > 0) {
+            $solution->getFromDB($_REQUEST['id']);
          }
+         $solution->showForm(null, $sol_params);
+      } else if (isset($_REQUEST[$parent->getForeignKeyField()])
+            && isset($_REQUEST["id"])
+            && $parent->getFromDB($_REQUEST[$parent->getForeignKeyField()])) {
+
+         $ol = ObjectLock::isLocked( $_REQUEST['parenttype'], $parent->getID() );
+         if ($ol && (Session::getLoginUserID() != $ol->fields['users_id'])) {
+            ObjectLock::setReadOnlyProfile( );
+         }
+
+         Ticket::showSubForm($item, $_REQUEST["id"], ['parent' => $parent,
+                                                      'tickets_id' => $_REQUEST["tickets_id"]]);
+      } else {
+         echo __('Access denied');
       }
+
       Html::ajaxFooter();
       break;
 }